Added nRF24_TransmitPacket function.
I have added the nRF24_TransmitPacket function in the library in order for it to be more complete and in order to provide the user a better ready made functionality, so the user does not need to implement the function. Also alongside with the addition of the function, I have added some necessary components in the header file. The function I added was written in the demo file provided, and I made some minor changes.