Added the dynamic payload functions.
Added the functions to enable or disable the dynamic payload feature of the NRF24 and also modified the receiving function to account for the dynamic payload. The RX buffer flushing that is included in the receive function, it will be removed in the future, and it will be replaced by an interrupt using the IRQ pin. Flushing is needed, because the bytes from the RX buffer are not fully emptied with dynamic payload, which may also be my mistake somewhere.